requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-multipart
jinja2
openpyxl
itsdangerous
redis
